backend/requirements.txt
# FastAPI Backend
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Market Data & Trading
alpaca-py==0.15.0
alpaca-trade-api==3.1.1

# Database
supabase==2.3.0
postgrest==0.13.0

# Cache
redis==5.0.1
upstash-redis==0.15.0

# Math & Scientific Computing
numpy==1.26.2
scipy==1.11.4
pandas==2.1.3

# AI/LLM
anthropic==0.7.8

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3

# HTTP Clients
httpx==0.25.2
websockets==12.0

# Logging & Monitoring
structlog==23.2.0

